pyproject.toml
[project]
name = "stableagent-backend"
version = "0.1.0"
description = "StableAgent agentic orchestration"
license = "MIT AND (Apache-2.0 OR BSD-2-Clause)"
readme = "README.md"
requires-python = ">=3.10.0"
dependencies = [
    "e2b-code-interpreter>=2.4.1",
    "fastapi>=0.128.0",
    "google-cloud-aiplatform>=1.132.0",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-anthropic>=1.3.0",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.6",
    "langchain-exa>=1.0.0",
    "langchain-google-community>=3.0.2",
    "langchain-google-genai>=4.1.2",
    "langchain-google-vertexai>=3.2.0",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.16",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-supervisor>=0.0.31",
    "privy-client>=0.5.0",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=22.0.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlparse>=0.5.5",
    "uvicorn>=0.40.0",
    # File processing dependencies for chat attachments
    "google-cloud-storage>=3.7.0",
    "pypdf>=6.5.0",
    "pandas>=2.3.3",
    "openpyxl>=3.1.5",
]
